Move appendIfMissing function to common location.
The appendIfMissing function is used by several physical backends so
move the function to a common file.
